Moved all array operation related functions to its corresponding class (ArrayOperations,.cs) and implemented IsArrayNullOrEmpty().
